+++ app/(tabs)/myPage/index.tsx
@@ -9,8 +9,8 @@ import BookmarkIcon from "@/assets/icons/bookmark-icon.svg";
 import LastTripIcon from "@/assets/icons/last-trip-icon.svg";
 import MyStoryIcon from "@/assets/icons/mystory-icon.svg";
 
+import { useBookmarkStore } from "@/store/common/useBookmarkStore";
 import { useAuthStore } from "@/store/profile/useAuthStore";
-import { useBookmarkStore } from "@/store/useBookmarkStore";
 
 export default function MyPageIndex() {
   const router = useRouter();
