+++ src/components/bottomSheet/RouteBottomSheet.tsx
@@ -12,9 +12,9 @@ import MyRouteSightList from "@/components/myRoute/sight/MyRouteSightList";
 
 import { theme } from "@/styles/theme";
 
+import { useBottomSheetStore } from "@/store/common/useBottomSheetStore";
 import { useMyRouteBottomSheetStore } from "@/store/route/useMyRouteBottomSheetStore";
 import { useRouteCartStore } from "@/store/route/useRouteCartStore";
-import { useBottomSheetStore } from "@/store/useBottomSheetStore";
 
 const RouteBottomSheet = () => {
   const routeCartItems = useRouteCartStore((state) => state.routeCartItems);
