+++ src/components/map/RouteMap.tsx
@@ -14,14 +14,14 @@ import Animated, {
 
 import { Ionicons } from "@expo/vector-icons";
 
-import { useLocation } from "@/hooks/useLocation";
+import { useLocation } from "@/hooks/common/useLocation";
 
 import { MapRef } from "@/types/map";
 import { SightInfo } from "@/types/sight";
 
 import { theme } from "@/styles/theme";
 
-import {useLocationStore} from "@/store/useLocationStore";
+import { useLocationStore } from "@/store/common/useLocationStore";
 
 const LOCATION_BUTTON_SIZE = 48;
 const LOCATION_BUTTON_MARGIN = 16;
@@ -45,10 +45,10 @@ const AnimatedTouchable = Animated.createAnimatedComponent(TouchableOpacity);
 
 const RouteMap = forwardRef<MapRef, RouteMapProps>(function RouteMapComponent(
   { animatedPosition, markers, children },
-  ref,
+  ref
 ) {
   const mapRef = useRef<MapView>(null);
-  const location = useLocationStore(state => state.location);
+  const location = useLocationStore((state) => state.location);
 
   useImperativeHandle(ref, () => ({
     getBoundaries: async () => {
@@ -70,7 +70,10 @@ const RouteMap = forwardRef<MapRef, RouteMapProps>(function RouteMapComponent(
   }));
 
   const moveToCurrentLocation = useCallback(async () => {
-    mapRef.current?.animateToRegion({...location, latitudeDelta: 0.01, longitudeDelta: 0.01}, 300);
+    mapRef.current?.animateToRegion(
+      { ...location, latitudeDelta: 0.01, longitudeDelta: 0.01 },
+      300
+    );
   }, [location]);
 
   const buttonAnimatedStyle = useAnimatedStyle(() => {
