+++ src/store/common/useLocationStore.ts
